Proof of Concept: Allow N-R19+ devices to accept the Xposed Framework
As this is a proof of concept, doing this would natively allow Zygote
to accept the XposedBridge.jar as a whitelisted Zygote initialized
binary.

DO NOT MERGE: Even though this is found on the Substratum code review, this
has no relation to anything OMS or Substratum, but just a proof of
concept to explain the whitelisting commit.

Change-Id: I69d33dcec1c755b18b7fc6ef6ea01993ca6e3bc5